pyproject.toml
[tool.poetry]
name = "tech-test"
version = "0.1.0"
description = ""
authors = ["Liam Garrison <liam@fluenthq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.27.0"
tabulate = "^0.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
